creating a way for user to add orders to a database

+++ src/main/java/domain/orders/Order.java
@@ -4,8 +4,8 @@ public class Order {
 
     private final int orderID;
     private final String orderStatus;
-    private final double length;
-    private final double width;
+    private final int length;
+    private final int width;
     private final String customerPhone;
     private final String customerEmail;
     private final double price;
@@ -13,7 +13,7 @@ public class Order {
 
 
 
-    public Order(int orderID, String orderStatus, double length, double width, String customerPhone, String customerEmail, double price, int salesmanID) {
+    public Order(int orderID, String orderStatus, int length, int width, String customerPhone, String customerEmail, double price, int salesmanID) {
         this.orderID = orderID;
         this.orderStatus = orderStatus;
         this.length = length;
